+++ CHANGELOG.md
@@ -6,10 +6,13 @@
 
 - Added `Finalize` support to `Process` interface. [#5](https://github.com/go-nacelle/process/pull/5)
 - Added `WithInitializerContextFilter` and `WithProcessContextFilter`. [#5](https://github.com/go-nacelle/process/pull/5)
+- Added `ConfigurationRegistry` and `Configurable` interfaces. [#7](https://github.com/go-nacelle/process/pull/7)
+- Added `LoadConfig` and `ValidateConfig` to the `Runner` interface. [#7](https://github.com/go-nacelle/process/pull/7)
 
 ### Changed
 
 - Added context parameters to `Init`, `Start`, `Stop`, and `Finalize` methods. [#5](https://github.com/go-nacelle/process/pull/5)
+- Removed config parameters from `Init` methods. [#7](https://github.com/go-nacelle/process/pull/7)
 
 ## [v1.1.0] - 2020-10-03